Move file to match global hierarchy
