+++ app/src/main/java/com/github/se/eduverse/repository/FolderRepositoryImpl.kt
@@ -1,5 +1,8 @@
-package com.github.se.eduverse.model.folder
+package com.github.se.eduverse.repository
 
+import com.github.se.eduverse.model.folder.FilterTypes
+import com.github.se.eduverse.model.folder.Folder
+import com.github.se.eduverse.model.folder.MyFile
 import com.google.firebase.firestore.DocumentSnapshot
 import com.google.firebase.firestore.FirebaseFirestore
 import java.io.InvalidClassException
