+++ app/src/main/java/com/github/se/eduverse/viewmodel/FolderViewModel.kt
@@ -1,8 +1,13 @@
-package com.github.se.eduverse.model.folder
+package com.github.se.eduverse.viewmodel
 
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
+import com.github.se.eduverse.model.folder.FilterTypes
+import com.github.se.eduverse.model.folder.Folder
+import com.github.se.eduverse.model.folder.MyFile
+import com.github.se.eduverse.repository.FolderRepository
+import com.github.se.eduverse.repository.FolderRepositoryImpl
 import com.google.firebase.Firebase
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.FirebaseUser
